Added logger infrastructure
This logger provides 5 levels of logs
    1. avd_log_debug(format string, args)
    2. avd_log_info(format string, args)
    3. avd_log_warn(format string, args)
    4. avd_log_error(format string, args)
    5. avd_log_fatal(format string, args)

This logger prints logs of format <TIME LEVEL FILE:LINE: MESSAGE>

Also moved from dynamic allocation to static memory allocation.
This wont cause much difference for now as the number of users and
supported workers are limited. This needs to be chaged to synamic
allocation in future for scaling.

When compiling for testing purpose do not use the optimization flag
as this might reorder the instructions. This will cause difficulty
while debugging using files and linenumbers.