Message marshalling
Marshall the data of different types into char buffer.
Currently basic c data types are supported.

Create custom message structs for different type of messages. These
messages will be marshalled in the message_t struct. message_t struct
comprises of two fields. Header and data. In every read first the header
will be read and then the exact bytes of the body will be read. This
will reduce the number of messages and also the network utlization.

Structure of message_t struct
    message_t
        - header
            -type
            -seq no
            -size
        - data
            -buf

We are adding seq no as currently the maximum buffer size is set at 1024
bytes. When sending shared libraries to the server and further to the
workers this buffer will fall short and thus we will be required to send
the file in multiple chunks. Seq no will help us in rebulding the file
correctly from the chunks.