Chunk Based File Transfer
There are 3 types of files involved in the system
    1. Task File - Shared library provided by the user
    2. Input File - This can be initial input by the user
                    or intermediate input provided to the
                    next worker by the current worker.
    3. Output File - It can be the final output being sent
                     to the server or intermediate output
                     generated by the workers.

Size of these files can be as large as 10-15 Mb. In such scenario
it is best to send the files in small chunks thus reducing the
size of message and also utilize the generic message passing
buffer which has capacity of 2048 bytes. That is 2Kb plus the size
of the header. The file being sent in broken down into chunks of
MAX_BUF_SZ and encapsulated with the message header along with the
appropriate flag, number of bytes sent and a sequence number. Seq
number helps in indentifying a missed file chunk. A finish flag is
sent along with the last message of the file. On receiving the finish
flag is it is a task file the file is given executing permissions and
the seq number for the file sending user is set to 1.