Task Session in User and Server
Maintaining server session object to reduce the number of operations
while performing search and retreival.

Update the task session objects while sending and receiving the tasks
from the user. A user directory will be maintained for each user which
will store all the files related to the task such as shared library, the
input files and the output files.

Update the cJSON library to latest version.